Plasma clinical analytes

Common clinical analytes measured in plasma. There are measurements for 10--20 rats per experimental group.

ANALYTES

Format

An object of class data.frame with 282 rows and 20 columns:

pid

character; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

iowa_id

character; unique ID used for each rat at the Iowa PASS.

omics_analysis

logical; whether the sample was selected for -omics analysis.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

group

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

runseq

integer; run order for 6M animals (1--146).

corticosterone

integer; corticosterone (ng/mL).

glucagon

numeric; glucagon (pM).

glucose

integer; glucose (mg/dL).

glycerol

numeric; glycerol (mg/dL).

insulin

integer; insulin (pg/mL).

insulin_iu

numeric; insulin (\(\mu\)IU/mL) using the conversion 1 pg/mL = 0.023 \(\mu\)IU/mL provided in the package insert.

insulin_pm

numeric; insulin (pM).

ketones

integer; total ketone bodies (\(\mu\)mol/L).

lactate

numeric; lactate (mmol/L).

leptin

integer; leptin (pg/mL).

nefa

numeric; non-esterified fatty acids (NEFA) (mmol/L).

Details

Analyte	Catalogue No.	Company

Corticosterone	55-CORMS-E01	Alpco (Salem, NH)

Glucose	B24985	Beckman Coulter (Brea, CA)

Glycerol	445850	Beckman Coulter

Lactate	A95550	Beckman Coulter

Total Ketones	415-73301, 411-73401	Fujifilm Wako (Osaka, Japan)

NEFA	995-34791, 999-34691, 993-35191, 991-34891	Fujifilm Wako

Glucagon	K1535YK	Meso Scale Discovery (Rockville, MD)

Insulin, Leptin	K15158C (multiplex assay)	Meso Scale Discovery

See also

Examples

head(ANALYTES)
#>        pid iowa_id omics_analysis age    sex group runseq corticosterone
#> 1 10043527  06F8T1           TRUE  6M Female   SED     81             47
#> 2 10043799  06F8T3          FALSE  6M Female   SED     NA             NA
#> 3 10043950  06F8T4           TRUE  6M Female   SED    121             38
#> 4 10044094  06F8T5          FALSE  6M Female   SED    103             43
#> 5 10044337  06F8T8           TRUE  6M Female   SED     42             17
#> 6 10044418  06F8T9          FALSE  6M Female   SED    122             NA
#>   glucagon glucose glycerol insulin insulin_iu insulin_pm ketones lactate
#> 1    145.7     147     1.29    2249     51.727      387.8     803    1.80
#> 2       NA      NA       NA      NA         NA         NA      NA      NA
#> 3     64.2     151     1.95    1655     38.065      285.3     917    1.38
#> 4     61.5      NA       NA     514     11.822       88.6     928      NA
#> 5    103.2     132     2.64    2332     53.636      402.1     996    1.23
#> 6     67.5      NA       NA    1778     40.894      306.6     694      NA
#>   leptin nefa
#> 1  14278 0.53
#> 2     NA   NA
#> 3  19476 1.00
#> 4  10055   NA
#> 5  18695 1.01
#> 6  17816   NA