Statistical analysis results of plasma clinical analytes

Statistical analysis results of plasma clinical analytes.

ANALYTES_STATS

Format

An object of class data.frame with 144 rows and 20 columns:

response

character; the clinical analyte being tested.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

contrast

factor; the comparison between groups. All contrasts are ratios between the trained and SED group means (levels "1W / SED", "2W / SED", "4W / SED", "8W / SED").

estimate_type

character; interpretation of the value in the estimate column. All "ratio" (ratio between group means, as specified by contrast).

null

numeric; the value of the estimate under the null hypothesis.

estimate

numeric; ratio between the means of the groups as specified by contrast.

SE

numeric; the standard error of the estimate.

lower.CL

numeric; lower bound of the 95% confidence interval.

upper.CL

numeric; upper bound of the 95% confidence interval.

statistic_type

character; the type of statistical test.

statistic

numeric; the value of the test statistic.

df

numeric; the number of residual degrees of freedom.

p.value

numeric; the p-value associated with the statistical test.

p.adj

numeric; the Dunnett-adjusted p-value.

signif

logical; TRUE if p.adj < 0.05.

model_type

character; the statistical model used. All "glm" (generalized linear model).

family

character; the probability distribution and link function for the generalized linear model.

formula

character; the model formula. Includes the response variable, any transformations, and predictors.

weights

character; if reciprocal group variances were used as weights to account for heteroscedasticity (nonconstant residual variance), this is noted here.

See also

Examples

head(ANALYTES_STATS)
#>         response age    sex contrast estimate_type null estimate        SE
#> 1 Corticosterone  6M Female 1W / SED         ratio    1 1.832279 0.2981685
#> 2 Corticosterone  6M Female 2W / SED         ratio    1 1.530332 0.2571677
#> 3 Corticosterone  6M Female 4W / SED         ratio    1 1.895274 0.2778634
#> 4 Corticosterone  6M Female 8W / SED         ratio    1 1.279948 0.2011438
#> 5 Corticosterone  6M   Male 1W / SED         ratio    1 1.832279 0.2981685
#> 6 Corticosterone  6M   Male 2W / SED         ratio    1 1.530332 0.2571677
#>    lower.CL upper.CL statistic_type statistic  df    p.value       p.adj signif
#> 1 1.2249426 2.740737              t  3.721236 189 0.06848359 0.001014465   TRUE
#> 2 1.0097111 2.319392              t  2.531939 189 0.12760657 0.043071377   TRUE
#> 3 1.3186265 2.724094              t  4.361023 189 0.05958248 0.000083877   TRUE
#> 4 0.8675887 1.888298              t  1.570596 189 0.20213238 0.333765251  FALSE
#> 5 1.2249426 2.740737              t  3.721236 189 0.06848359 0.001014465   TRUE
#> 6 1.0097111 2.319392              t  2.531939 189 0.12760657 0.043071377   TRUE
#>   model_type          family                      formula
#> 1        glm gaussian("log") corticosterone ~ sex + group
#> 2        glm gaussian("log") corticosterone ~ sex + group
#> 3        glm gaussian("log") corticosterone ~ sex + group
#> 4        glm gaussian("log") corticosterone ~ sex + group
#> 5        glm gaussian("log") corticosterone ~ sex + group
#> 6        glm gaussian("log") corticosterone ~ sex + group
#>                   weights
#> 1 inverse group variances
#> 2 inverse group variances
#> 3 inverse group variances
#> 4 inverse group variances
#> 5 inverse group variances
#> 6 inverse group variances