Statistical analysis results of baseline (pre-training) measures

Statistical analysis results of baseline (pre-training) measures: NMR mass, lean mass, fat mass, % lean mass, % fat mass; absolute VO\(_2\)max; relative VO\(_2\)max; and maximum run speed.

BASELINE_STATS

Format

An object of class data.frame with 112 rows and 23 columns:

response

character; the measurement being tested.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

contrast

character; the comparison between groups. Either ratios or differences between trained and sedentary group means.

estimate_type

character; interpretation of the value in the estimate column. Either "ratio" (ratio between means) or "location shift" (difference of signed mean ranks).

null

numeric; the value of the estimate under the null hypothesis.

estimate

numeric; ratio or difference between the means of the groups as specified by contrast.

SE

numeric; the standard error of the estimate.

lower.CL

numeric; lower bound of the 95% confidence interval.

upper.CL

numeric; upper bound of the 95% confidence interval.

statistic_type

character; the type of statistical test. Either "t" (Student's t-statistic), "z" (Z-statistic), or "W" (Wilcox's W-statistic).

statistic

numeric; the value of the test statistic.

n_SED

integer; the number of samples in the SED (sedentary) group.

n_trained

integer; the number of samples in the trained group (specified by contrast).

df

numeric; the number of residual degrees of freedom.

p.value

numeric; the p-value associated with the statistical test.

p.adj

numeric; the adjusted p-value.

signif

logical; TRUE if p.adj < 0.05.

model_type

character; the statistical model used for testing.

family

character; the probability distribution and link function for the generalized linear model.

formula

character; the model formula. Includes the response variable, any transformations, and predictors.

weights

character; if reciprocal group variances were used as weights to account for heteroscedasticity (nonconstant residual variance), this is noted here.

obs_removed

character; if any observations were omitted during the analysis, they are listed here.

See also

Examples

unique(BASELINE_STATS$response) # available measures
#> [1] "NMR Body Mass"     "NMR Lean Mass"     "NMR Fat Mass"     
#> [4] "NMR % Lean"        "NMR % Fat"         "Absolute VO2max"  
#> [7] "Relative VO2max"   "Maximum Run Speed"

head(BASELINE_STATS)
#>        response age    sex contrast estimate_type null  estimate         SE
#> 1 NMR Body Mass  6M Female 1W / SED         ratio    1 1.0602759 0.01362695
#> 2 NMR Body Mass  6M Female 2W / SED         ratio    1 1.0542059 0.01396689
#> 3 NMR Body Mass  6M Female 4W / SED         ratio    1 0.9932244 0.01480814
#> 4 NMR Body Mass  6M Female 8W / SED         ratio    1 1.0040924 0.01246170
#> 5 NMR Body Mass 18M Female 1W / SED         ratio    1 0.9447209 0.01387118
#> 6 NMR Body Mass 18M Female 2W / SED         ratio    1 0.9598096 0.01492037
#>    lower.CL  upper.CL statistic_type  statistic n_SED n_trained  df
#> 1 1.0271734 1.0944452              t  4.5539975    NA        NA 271
#> 2 1.0202941 1.0892448              t  3.9843646    NA        NA 271
#> 3 0.9573432 1.0304504              t -0.4560039    NA        NA 271
#> 4 0.9738041 1.0353228              t  0.3290707    NA        NA 271
#> 5 0.9111007 0.9795817              t -3.8729390    NA        NA 271
#> 6 0.9236847 0.9973474              t -2.6387900    NA        NA 271
#>        p.value        p.adj signif model_type          family
#> 1 7.961740e-06 3.156009e-05   TRUE        glm gaussian("log")
#> 2 8.702181e-05 3.412170e-04   TRUE        glm gaussian("log")
#> 3 1.351248e+00 9.446422e-01  FALSE        glm gaussian("log")
#> 4 7.423564e-01 9.743686e-01  FALSE        glm gaussian("log")
#> 5 1.999865e+00 5.271524e-04   TRUE        glm gaussian("log")
#> 6 1.991198e+00 3.168499e-02   TRUE        glm gaussian("log")
#>                                         formula                    weights
#> 1 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#> 2 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#> 3 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#> 4 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#> 5 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#> 6 pre_body_mass ~ age + group + sex + age:group reciprocal group variances
#>   obs_removed
#> 1        <NA>
#> 2        <NA>
#> 3        <NA>
#> 4        <NA>
#> 5        <NA>
#> 6        <NA>