Muscle fiber count multivariate t-test results

Muscle fiber count multivariate t-test results.

FIBER_COUNT_MVT

Format

An object of class data.frame with 16 rows and 17 columns:

response

character; the measure being tested. All "Fiber Count".

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

muscle

factor; the muscle that was sampled with levels "LG" (lateral gastrocnemius), "MG" (medial gastrocnemius), "PL" (plantaris), and "SOL" (soleus).

contrast

factor; the comparison between groups. All contrasts are ratios between the 8W and SED group means ("8W / SED").

estimate_type

character; interpretation of the value in the estimate column. All "ratio" (ratio between group means, as specified by contrast).

estimate

numeric; ratio between the means of the groups as specified by contrast.

SE

numeric; the standard error of the estimate. Only applicable when statistic_type = "t".

statistic_type

character; the type of statistical test.

statistic

numeric; the value of the test statistic.

df1, df2

numeric; the degrees of freedom. See help("FDist") for details. If statistic_type == "t", then df1 is not applicable.

p.value

numeric; the p-value associated with the statistical test.

p.adj

numeric; the Holm-adjusted p-value.

signif

logical; TRUE if p.adj < 0.05.

model_type

character; the statistical model used for testing.

formula

character; the model formula. Includes the response variable, any transformations, and predictors.