Statistical analysis results of muscle-specific measures

Statistical analysis results of muscle-specific measures: terminal muscle mass, mean cross-sectional area (CSA), glycogen, capillary contacts, and citrate synthase.

MUSCLES_STATS

Format

An object of class data.frame with 224 rows and 22 columns:

response

character; the muscle-specific measurement being tested.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

muscle

factor; the muscle that was sampled with levels "LG" (lateral gastrocnemius), "MG" (medial gastrocnemius), "PL" (plantaris), and "SOL" (soleus).

contrast

factor; the comparison between groups. All contrasts are ratios between the trained and SED group means (levels "1W / SED", "2W / SED", "4W / SED", "8W / SED").

estimate_type

character; interpretation of the value in the estimate column. All "ratio" (ratio between group means, as specified by contrast).

null

numeric; the value of the estimate under the null hypothesis.

estimate

numeric; ratio between the means of the groups as specified by contrast.

SE

numeric; the standard error of the estimate.

lower.CL

numeric; lower bound of the 95% confidence interval.

upper.CL

numeric; upper bound of the 95% confidence interval.

statistic_type

character; the type of statistical test. All "t" (Student's t-statistic).

statistic

numeric; the value of the test statistic.

df

numeric; the number of residual degrees of freedom.

p.value

numeric; the p-value associated with the statistical test.

p.adj

numeric; the adjusted p-value.

signif

logical; TRUE if p.adj < 0.05.

model_type

character; the statistical model used for testing. All "lme" (linear mixed effects model).

fixed

character; the fixed-effects component of the model formula. Describes the response variable (and any transformations) and all predictor variables that were included. Equivalent to formula.

random

character; the random component of the model formula. Specifies the random effects for mixed effects models. All "~1 | pid" (random intercept for each rat).

weights

character; if reciprocal group variances were used as weights to account for heteroscedasticity (nonconstant residual variance), this is noted here.

obs_removed

character; if any observations were omitted during the analysis, they are listed here.

See also

Examples

unique(MUSCLES_STATS$response) # available measures
#> [1] "Capillary Contacts"   "Citrate Synthase"     "Glycogen"            
#> [4] "Mean CSA"             "Terminal Muscle Mass"

head(MUSCLES_STATS)
#> # A tibble: 6 × 22
#>   response       age   sex   muscle contrast estimate_type  null estimate     SE
#>   <chr>          <fct> <fct> <fct>  <fct>    <chr>         <dbl>    <dbl>  <dbl>
#> 1 Capillary Con… 6M    Fema… LG     8W / SED ratio             1    1.05  0.0257
#> 2 Capillary Con… 6M    Fema… MG     8W / SED ratio             1    0.974 0.0401
#> 3 Capillary Con… 6M    Fema… PL     8W / SED ratio             1    0.948 0.0226
#> 4 Capillary Con… 6M    Fema… SOL    8W / SED ratio             1    1.05  0.0359
#> 5 Capillary Con… 6M    Male  LG     8W / SED ratio             1    1.05  0.0257
#> 6 Capillary Con… 6M    Male  MG     8W / SED ratio             1    0.974 0.0401
#> # ℹ 13 more variables: lower.CL <dbl>, upper.CL <dbl>, statistic_type <chr>,
#> #   statistic <dbl>, df <dbl>, p.value <dbl>, p.adj <dbl>, signif <lgl>,
#> #   model_type <chr>, fixed <chr>, random <chr>, weights <chr>,
#> #   obs_removed <chr>