NMR body composition measures

Measures of body composition (total body mass, body fat, % body fat, lean mass, and % lean mass) both pre- and post-training.

NMR

Format

An object of class data.frame with 282 rows and 19 columns:

pid

character; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

iowa_id

character; unique ID used for each rat at the Iowa PASS.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

group

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

pre_day

POSIXct, POSIXt; the day when the pre-training NMR body composition measures were recorded.

pre_body_mass

numeric; body mass (g) recorded on the same day as NMR body composition measures (pre-training).

pre_fat

numeric; fat mass (g) recorded via NMR (pre-training).

pre_lean

numeric; lean mass (g) recorded via NMR (pre-training).

pre_fat_pct

numeric; percent fat mass (pre-training). Calculated as pre_fat / pre_body_mass * 100.

pre_lean_pct

numeric; percent lean mass (pre-training). Calculated as pre_lean / pre_body_mass * 100.

pre_fluid_pct

numeric; percent fluid mass (pre-training).

post_day

POSIXct, POSIXt; the day when the post-training NMR body composition measures were recorded.

post_body_mass

numeric; body mass (g) recorded on the same day as NMR body composition measures (post-training).

post_fat

numeric; fat mass (g) recorded via NMR (post-training).

post_lean

numeric; lean mass (g) recorded via NMR (post-training).

post_fat_pct

numeric; percent fat mass (post-training). Calculated as post_fat / post_body_mass * 100.

post_lean_pct

numeric; percent lean mass (post-training). Calculated as post_lean / post_body_mass * 100.

post_fluid_pct

numeric; percent fluid mass (post-training).

See also

Examples

head(NMR)
#>        pid iowa_id age    sex group    pre_day pre_body_mass pre_fat pre_lean
#> 1 10043527  06F8T1  6M Female   SED 2018-07-25         181.2 20.2944 109.2636
#> 2 10043799  06F8T3  6M Female   SED 2018-07-25         179.0 19.1530 105.9680
#> 3 10043950  06F8T4  6M Female   SED 2018-07-25         174.4 19.1840 103.2448
#> 4 10044094  06F8T5  6M Female   SED 2018-07-25         181.2 20.4756 105.4584
#> 5 10044337  06F8T8  6M Female   SED 2018-07-25         185.2 21.2980 108.8976
#> 6 10044418  06F8T9  6M Female   SED 2018-07-25         178.3 20.1479 104.1272
#>   pre_fat_pct pre_lean_pct pre_fluid_pct   post_day post_body_mass post_fat
#> 1        11.2         60.3           9.5 2018-09-26            199  21.2004
#> 2        10.7         59.2           9.6 2018-09-30            203  25.4180
#> 3        11.0         59.2          10.0 2018-09-30            202  27.0320
#> 4        11.3         58.2           9.7 2018-09-26            190  24.6432
#> 5        11.5         58.8          10.1 2018-09-26            205  27.5948
#> 6        11.3         58.4           9.6 2018-09-27            191  24.4271
#>   post_lean post_fat_pct post_lean_pct post_fluid_pct
#> 1  107.9952         11.7          59.6           10.0
#> 2   99.7030         14.2          55.7            9.3
#> 3   96.9664         15.5          55.6            9.8
#> 4  103.8276         13.6          57.3           10.1
#> 5  101.4896         14.9          54.8            9.7
#> 6  102.7008         13.7          57.6            9.7