Statistical analysis results of post - pre training differences

Statistical analysis results of post - pre training differences: NMR body mass, lean mass, fat mass, % lean mass, % fat mass; difference between terminal and pre-training NMR mass at 1W and 2W of endurance training; absolute VO\(_2\)max; relative VO\(_2\)max; and maximum run speed.

PRE_POST_STATS

Format

An object of class data.frame with 90 rows and 21 columns:

response

character; the measurement being tested.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

group

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

estimate_type

character; interpretation of the value in the estimate column. Either "(post - pre) mean" (mean of the paired (post - pre) differences) or "signed mean (post - pre) rank" (the signed mean ranks of the (post - pre) differences).

estimate

numeric; the value of the estimate, as described by estimate_type.

SE

numeric; the standard error of the estimate.

lower.CL

numeric; lower bound of the 95% confidence interval.

upper.CL

numeric; upper bound of the 95% confidence interval.

statistic_type

character; the type of statistical test. Either "t" (Student's t-statistic) or "W" (Wilcox's W-statistic).

statistic

numeric; the value of the test statistic.

p.value

numeric; the p-value associated with the statistical test.

p.adj

numeric; the Holm-adjusted p-value.

signif

logical; TRUE if p.adj < 0.05.

model_type

character; the statistical model used for testing. Either "lm" (linear regression) or "wilcox.test" (Wilcoxon test).

formula

character; the model formula. Includes the response variable, any transformations, and predictors.

family

character; the probability distribution and link function for the generalized linear model.

weights

character; if reciprocal group variances were used as weights to account for heteroscedasticity (nonconstant residual variance), this is noted here.

obs_removed

character; if any observations were omitted during the analysis, they are listed here.

Examples

unique(PRE_POST_STATS$response)
#> [1] "Absolute VO2max"          "Maximum Run Speed"       
#> [3] "NMR % Fat"                "NMR % Lean"              
#> [5] "NMR Body Mass"            "NMR Fat Mass"            
#> [7] "NMR Lean Mass"            "Relative VO2max"         
#> [9] "Term - NMR Pre Body Mass"

head(PRE_POST_STATS)
#> # A tibble: 6 × 20
#>   response      age   sex   group estimate_type estimate    SE lower.CL upper.CL
#>   <chr>         <fct> <fct> <fct> <chr>            <dbl> <dbl>    <dbl>    <dbl>
#> 1 Absolute VO2… 6M    Fema… SED   (post - pre)…   0.131  0.339  -0.539     0.802
#> 2 Absolute VO2… 6M    Fema… 4W    (post - pre)…   0.504  0.263  -0.0157    1.02 
#> 3 Absolute VO2… 6M    Fema… 8W    (post - pre)…   2.59   0.285   2.03      3.16 
#> 4 Absolute VO2… 6M    Male  SED   (post - pre)…  -0.0677 0.339  -0.738     0.603
#> 5 Absolute VO2… 6M    Male  4W    (post - pre)…   0.796  0.277   0.248     1.34 
#> 6 Absolute VO2… 6M    Male  8W    (post - pre)…   2.50   0.326   1.86      3.14 
#> # ℹ 11 more variables: statistic_type <chr>, statistic <dbl>, df <dbl>,
#> #   n <int>, p.value <dbl>, p.adj <dbl>, signif <lgl>, model_type <chr>,
#> #   formula <chr>, weights <chr>, obs_removed <chr>