VO2max test measures

Measures from the pre- and post-training VO\(_2\)max tests: total body mass, absolute and relative VO\(_2\)max, respiratory exchange rate (RER), VCO\(_2\), maximum run speed, blood lactate, and test duration.

VO2MAX

Format

An object of class data.frame with 234 rows and 27 columns:

pid

character; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

iowa_id

character; unique ID used for each rat at the Iowa PASS.

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

sex

factor; the sex of the rat with levels "Female" and "Male".

group

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

pre_body_mass

numeric; body mass (g) recorded on the same day as the VO\(_2\)max test (pre-training).

pre_t_start

POSIXct, POSIXt; date and time of the VO\(_2\)max test start (pre-training).

pre_blactate_begin

numeric; blood lactate (mmol/L) recorded at the start of the test (pre-training).

pre_vo2max_ml_kg_min

numeric; maximum relative VO\(_2\)max (mL/kg body mass/min) reached during the test (pre-training).

pre_vo2max_ml_min

numeric; maximum absolute VO\(_2\)max (mL/kg/min) reached during the test (pre-training).

pre_vco2_max

numeric; maximum VCO\(_2\) reached during the VO\(_2\)max test (pre-training).

pre_rer_max

numeric; maximum respiratory exchange rate (RER) reached during the VO\(_2\)max test (pre-training).

pre_speed_max

numeric; maximum run speed (m/min) reached during the VO\(_2\)max test (pre-training).

pre_t_complete

POSIXct, POSIXt; date and time of the VO\(_2\)max test end (pre-training).

pre_blactate_end

numeric; blood lactate (mmol/L) recorded at the end of the test (pre-training).

pre_comments

character; comments regarding the pre-training VO\(_2\)max test.

post_body_mass

numeric; body mass (g) recorded on the same day as the VO\(_2\)max test (post-training).

post_t_start

POSIXct, POSIXt; date and time of the VO\(_2\)max test start (post-training).

post_blactate_begin

numeric; blood lactate (mmol/L) recorded at the start of the test (post-training).

post_vo2max_ml_kg_min

numeric; maximum relative VO\(_2\)max (mL/kg body mass/min) reached during the test (post-training).

post_vo2max_ml_min

numeric; maximum absolute VO\(_2\)max (mL/kg/min) reached during the test (post-training).

post_vco2_max

numeric; maximum VCO\(_2\) reached during the VO\(_2\)max test (post-training).

post_rer_max

numeric; maximum respiratory exchange rate (RER) reached during the VO\(_2\)max test (post-training).

post_speed_max

numeric; maximum run speed (m/min) reached during the VO\(_2\)max test (post-training).

post_t_complete

POSIXct, POSIXt; date and time of the VO\(_2\)max test end (post-training).

post_blactate_end

numeric; blood lactate (mmol/L) recorded at the end of the test (post-training).

post_comments

character; comments regarding the post-training VO\(_2\)max test.

Examples

head(VO2MAX)
#>        pid iowa_id age    sex group pre_body_mass         pre_t_start
#> 1 10043527  06F8T1  6M Female   SED         178.3 2018-07-30 10:03:44
#> 2 10043799  06F8T3  6M Female   SED         182.8 2018-07-30 14:01:35
#> 3 10043950  06F8T4  6M Female   SED         178.2 2018-07-30 15:04:29
#> 4 10044094  06F8T5  6M Female   SED         182.0 2018-07-30 16:12:44
#> 5 10044337  06F8T8  6M Female   SED         182.4 2018-07-31 10:09:26
#> 6 10044418  06F8T9  6M Female   SED         174.5 2018-07-31 12:14:06
#>   pre_blactate_begin pre_vo2max_ml_kg_min pre_vo2max_ml_min pre_vco2_max
#> 1                1.9                 83.8          14.94154        79.72
#> 2                1.8                 71.7          13.10676        75.00
#> 3                2.3                 72.5          12.91950        76.50
#> 4                2.2                 69.7          12.68540        71.08
#> 5                2.1                 69.6          12.69504        76.19
#> 6                1.6                 77.6          13.54120        76.09
#>   pre_rer_max pre_speed_max      pre_t_complete pre_blactate_end
#> 1        1.03          25.2 2018-07-30 10:39:54              8.1
#> 2        1.08          25.2 2018-07-30 14:37:45             11.0
#> 3        1.06          25.2 2018-07-30 15:38:49              9.4
#> 4        1.04          25.2 2018-07-30 16:47:04              4.9
#> 5        1.10          23.4 2018-07-31 10:42:56             12.6
#> 6        1.00          23.4 2018-07-31 12:58:11             11.8
#>                                                                pre_comments
#> 1                                                                      <NA>
#> 2                                                                      <NA>
#> 3                                                                      <NA>
#> 4                                     Sat on shock too long towards the end
#> 5                                                                      <NA>
#> 6 Nosebleed, potentially from sitting on the shock too much towards the end
#>   post_body_mass        post_t_start post_blactate_begin post_vo2max_ml_kg_min
#> 1          199.2 2018-09-25 12:30:00                 1.8                 73.16
#> 2          203.0 2018-09-29 17:33:00                 1.9                 63.77
#> 3          202.3 2018-09-29 18:24:00                 1.6                 64.63
#> 4          190.4 2018-09-25 15:45:00                 1.9                 61.85
#> 5          204.6 2018-09-25 13:10:00                 1.1                 73.52
#> 6          190.8 2018-09-26 17:35:00                 1.0                 63.65
#>   post_vo2max_ml_min post_vco2_max post_rer_max post_speed_max
#> 1           14.57347         80.68         1.09           23.4
#> 2           12.94531         69.09         1.11           21.6
#> 3           13.07465         67.41         1.06           21.6
#> 4           11.77624         63.09         1.03           19.8
#> 5           15.04219         74.19         1.01           23.4
#> 6           12.14442         69.17         1.09           21.6
#>       post_t_complete post_blactate_end post_comments
#> 1 2018-09-25 13:13:00              12.2          <NA>
#> 2 2018-09-29 18:17:00              10.7          <NA>
#> 3 2018-09-29 19:08:00              11.2          <NA>
#> 4 2018-09-25 16:29:00               7.6          <NA>
#> 5 2018-09-25 13:56:00               9.0          <NA>
#> 6 2018-09-26 18:23:00              10.1          <NA>