Weekly body mass and blood lactate measures

Weekly measures of total body mass and blood lactate.

WEEKLY_MEASURES

Format

An object of class data.frame with 2226 rows and 11 columns:

pid

character; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

iowa_id

character; unique ID used for each rat at the Iowa PASS.

omics_analysis

logical; whether the sample was selected for -omics analysis.

sex

factor; the sex of the rat with levels "Female" and "Male".

group

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

age

factor; the age of the rat at the beginning of the training protocol. Levels: "6M" (Adult) and "18M" (Aged).

day

numeric; the day the measurement was recorded relative to the start of the training protocol.

week

numeric; the week the measurement was recorded relative to the start of the training protocol (1--8).

week_time

factor; at what point in the week the measurement was taken ("start" or "end").

body_mass

numeric; total body mass (g) recorded at the start of each week.

lactate

numeric; blood lactate (mM) recorded at the start and end of each week.

See also

Examples

head(WEEKLY_MEASURES)
#> # A tibble: 6 × 11
#>   pid   iowa_id omics_analysis sex   group age     day  week week_time body_mass
#>   <chr> <chr>   <lgl>          <fct> <fct> <fct> <dbl> <dbl> <fct>         <dbl>
#> 1 1004… 06F8T1  TRUE           Fema… SED   6M        1     1 start           181
#> 2 1004… 06F8T3  FALSE          Fema… SED   6M        1     1 start           182
#> 3 1004… 06F8T4  TRUE           Fema… SED   6M        1     1 start           176
#> 4 1004… 06F8T5  FALSE          Fema… SED   6M        1     1 start           182
#> 5 1004… 06F8T8  TRUE           Fema… SED   6M        1     1 start           187
#> 6 1004… 06F8T9  FALSE          Fema… SED   6M        1     1 start           179
#> # ℹ 1 more variable: lactate <dbl>